vc2013 -> vc2015

+++ src/Preferences.cpp
@@ -89,8 +89,8 @@ int Preferences::addCinderVersionInst( const QString &name, const QString &path,
 {
 	// see if we already have this version
 	for( size_t idx = 0; idx < mCinderVersions.size(); ++idx ) {
-		if( QFileInfo( path ).absoluteFilePath() == QFileInfo( mCinderVersions[idx].path ).absoluteFilePath() )
-			return idx;
+        if( QFileInfo( path ).absoluteFilePath() == QFileInfo( mCinderVersions[(int)idx].path ).absoluteFilePath() )
+            return (int)idx;
 	}
 
 	CinderVersion v;
@@ -113,7 +113,7 @@ void Preferences::removeCinderVersionInst( size_t index )
 
 void Preferences::updateCinderVersionInst( size_t index, const QString &name, const QString &path )
 {
-	Preferences::CinderVersion &cv( getCinderVersions()[index] );
+    Preferences::CinderVersion &cv( getCinderVersions()[(int)index] );
 	cv.name = name;
 	cv.path = path;
